+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -8,7 +8,7 @@ import sinon from 'sinon';
 import * as internalMediaModule from '@webex/internal-media-core';
 import StateMachine from 'javascript-state-machine';
 import uuid from 'uuid';
-import {assert} from '@webex/test-helper-chai';
+import {assert, expect} from '@webex/test-helper-chai';
 import {Credentials, Token, WebexPlugin} from '@webex/webex-core';
 import Support from '@webex/internal-plugin-support';
 import MockWebex from '@webex/test-helper-mock-webex';
@@ -2722,6 +2722,7 @@ describe('plugin-meetings', () => {
               createSendSlot: sinon.stub().returns({
                 publishStream: sinon.stub(),
                 unpublishStream: sinon.stub(),
+                setNamedMediaGroups: sinon.stub(),
               }),
               enableMultistreamAudio: sinon.stub(),
             };
@@ -5772,6 +5773,7 @@ describe('plugin-meetings', () => {
           meeting.mediaProperties.webrtcMediaConnection = {};
           meeting.audio = {handleLocalStreamChange: sinon.stub()};
           meeting.video = {handleLocalStreamChange: sinon.stub()};
+          meeting.statsAnalyzer = {updateMediaStatus: sinon.stub()};
           fakeMultistreamRoapMediaConnection = {
             createSendSlot: () => {
               return {
@@ -5838,6 +5840,10 @@ describe('plugin-meetings', () => {
               options: {meetingId: meeting.id},
             });
             assert.equal(meeting.mediaProperties.mediaDirection.sendShare, true);
+
+            assert.calledWith(meeting.statsAnalyzer.updateMediaStatus, {
+              expected: {sendShare: true},
+            });
           };
 
           const checkScreenShareAudioPublished = (stream) => {
@@ -5854,6 +5860,10 @@ describe('plugin-meetings', () => {
             );
             assert.equal(meeting.mediaProperties.shareAudioStream, stream);
             assert.equal(meeting.mediaProperties.mediaDirection.sendShare, true);
+
+            assert.calledWith(meeting.statsAnalyzer.updateMediaStatus, {
+              expected: {sendShare: true},
+            });
           };
 
           it('requests screen share floor and publishes the screen share video stream', async () => {
@@ -5948,6 +5958,11 @@ describe('plugin-meetings', () => {
 
             assert.equal(meeting.mediaProperties.shareVideoStream, null);
             assert.equal(meeting.mediaProperties.mediaDirection.sendShare, shareDirection);
+            if (!shareDirection) {
+              assert.calledWith(meeting.statsAnalyzer.updateMediaStatus, {
+                expected: {sendShare: false},
+              });
+            }
           };
 
           // share direction will remain true if only one of the two share streams are unpublished
@@ -5960,6 +5975,11 @@ describe('plugin-meetings', () => {
 
             assert.equal(meeting.mediaProperties.shareAudioStream, null);
             assert.equal(meeting.mediaProperties.mediaDirection.sendShare, shareDirection);
+            if (!shareDirection) {
+              assert.calledWith(meeting.statsAnalyzer.updateMediaStatus, {
+                expected: {sendShare: false},
+              });
+            }
           };
 
           it('fails if there is no media connection', async () => {
@@ -6029,6 +6049,29 @@ describe('plugin-meetings', () => {
             assert.notCalled(meeting.releaseScreenShareFloor);
           });
         });
+      });
+    });
+
+    describe('#setSendNamedMediaGroup', () => {
+      beforeEach(() => {
+        meeting.sendSlotManager.setNamedMediaGroups = sinon.stub().returns(undefined);
+      });
+      it('should throw error if not audio type', () => {
+        expect(() => meeting.setSendNamedMediaGroup(MediaType.VideoMain, 20)).to.throw(`cannot set send named media group which media type is ${MediaType.VideoMain}`)
+
+      });
+      it('fails if there is no media connection', () => {
+
+        meeting.mediaProperties.webrtcMediaConnection = undefined;
+        meeting.setSendNamedMediaGroup('AUDIO-MAIN', 20);
+        assert.notCalled(meeting.sendSlotManager.setNamedMediaGroups);
+      });
+
+      it('success if there is media connection', () => {
+        meeting.isMultistream = true;
+        meeting.mediaProperties.webrtcMediaConnection = true;
+        meeting.setSendNamedMediaGroup("AUDIO-MAIN", 20);
+        assert.calledOnceWithExactly(meeting.sendSlotManager.setNamedMediaGroups, "AUDIO-MAIN", [{type: 1, value: 20}]);
       });
     });
 
@@ -8172,7 +8215,7 @@ describe('plugin-meetings', () => {
               function: 'setUpLocusInfoMeetingInfoListener',
             },
             'meeting:meetingInfoUpdated'
-          )
+          );
 
           callback({isIntialized: false});
 
@@ -8190,7 +8233,7 @@ describe('plugin-meetings', () => {
               function: 'setUpLocusInfoMeetingInfoListener',
             },
             'meeting:meetingInfoUpdated'
-          )
+          );
         });
       });